Unified number conversion code
